Moving all non top-level code files into the src directory and fixing imports
