--- mouse-tracking-runtime/pytorch_inference/__init__.py
@@ -1,3 +1,0 @@
-from .single_pose import infer_single_pose_pytorch
-from .multi_pose import infer_multi_pose_pytorch
-from .fecal_boli import infer_fecal_boli_pytorch
