+++ pyproject.toml
@@ -5,11 +5,13 @@ description = "Runtime environment for mouse tracking experiments"
 requires-python = ">=3.10"
 packages = ["src/mouse_tracking"]
 dependencies = [
+    "absl-py>=2.3.0",
     "click==8.1.8",
     "contourpy==1.3.2",
     "cycler==0.12.1",
     "fonttools==4.57.0",
     "h5py==3.13.0",
+    "imageio>=2.37.0",
     "kiwisolver==1.4.8",
     "matplotlib==3.10.1",
     "mypy-extensions==1.0.0",
@@ -28,6 +30,7 @@ dependencies = [
     "pytz==2025.1",
     "scipy==1.15.2",
     "six==1.17.0",
+    "tensorflow==2.14",
     "torch>=2.7.1",
     "typer>=0.16.0",
     "tzdata==2025.1",
