+++ src/mouse_tracking/models/model_definitions.py
