+++ src/mouse_tracking/pytorch_inference/hrnet/config/__init__.py
