+++ src/mouse_tracking/pytorch_inference/hrnet/config/default.py
