+++ src/mouse_tracking/pytorch_inference/hrnet/config/models.py
