+++ src/mouse_tracking/pytorch_inference/single_pose.py
@@ -4,14 +4,15 @@ import numpy as np
 import queue
 import time
 import sys
-from utils.pose import render_pose_overlay
-from utils.hrnet import argmax_2d_torch, preprocess_hrnet
-from utils.prediction_saver import prediction_saver
-from utils.writers import write_pose_v2_data
-from utils.timers import time_accumulator
-from models.model_definitions import SINGLE_MOUSE_POSE
+from mouse_tracking.utils.pose import render_pose_overlay
+from mouse_tracking.utils.hrnet import argmax_2d_torch, preprocess_hrnet
+from mouse_tracking.utils.prediction_saver import prediction_saver
+from mouse_tracking.utils.writers import write_pose_v2_data
+from mouse_tracking.utils.timers import time_accumulator
+from mouse_tracking.models.model_definitions import SINGLE_MOUSE_POSE
 import torch
 import torch.backends.cudnn as cudnn
+# TODO: Where is this import file?
 from .hrnet.models import pose_hrnet
 from .hrnet.config import cfg
