+++ src/mouse_tracking/tfs_inference/food_hopper.py
@@ -6,11 +6,11 @@ import cv2
 import queue
 import time
 import sys
-from utils.static_objects import filter_static_keypoints, plot_keypoints, get_mask_corners
-from utils.prediction_saver import prediction_saver
-from utils.writers import write_static_object_data
-from utils.timers import time_accumulator
-from models.model_definitions import STATIC_FOOD_CORNERS
+from mouse_tracking.utils.static_objects import filter_static_keypoints, plot_keypoints, get_mask_corners
+from mouse_tracking.utils.prediction_saver import prediction_saver
+from mouse_tracking.utils.writers import write_static_object_data
+from mouse_tracking.utils.timers import time_accumulator
+from mouse_tracking.models.model_definitions import STATIC_FOOD_CORNERS
 
 
 def infer_food_hopper_model(args):
