+++ src/mouse_tracking/tfs_inference/multi_segmentation.py
@@ -5,11 +5,11 @@ import numpy as np
 import queue
 import time
 import sys
-from utils.segmentation import get_contours, pad_contours, render_segmentation_overlay, merge_multiple_seg_instances
-from utils.prediction_saver import prediction_saver
-from utils.writers import write_seg_data
-from utils.timers import time_accumulator
-from models.model_definitions import MULTI_MOUSE_SEGMENTATION
+from mouse_tracking.utils.segmentation import get_contours, pad_contours, render_segmentation_overlay, merge_multiple_seg_instances
+from mouse_tracking.utils.prediction_saver import prediction_saver
+from mouse_tracking.utils.writers import write_seg_data
+from mouse_tracking.utils.timers import time_accumulator
+from mouse_tracking.models.model_definitions import MULTI_MOUSE_SEGMENTATION
 from absl import logging
