+++ src/mouse_tracking/utils/segmentation.py
