+++ src/mouse_tracking/utils/static_objects.py
